+++ Sources/Models/StatusDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  StatusDelegate.swift
+//  TezosSwift
+//
+//  Created by Marek Fořt on 11/8/18.
+//  Copyright © 2018 Keefer Taylor. All rights reserved.
+//
+
+import Foundation
+
+public struct StatusDelegate: Codable {
+    let setable: Bool
+    let value: String?
+}
